+++ taoq-boot-common/taoq-boot-starter-system/src/main/java/tech/taoq/system/service/impl/DictTypeServiceImpl.java
@@ -4,9 +4,10 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import tech.taoq.common.exception.client.ClientErrorException;
 import tech.taoq.common.exception.client.ParamIllegalException;
 import tech.taoq.common.pojo.PageDto;
-import tech.taoq.system.domain.db.DictTypeDO;
+import tech.taoq.system.domain.DictTypeDO;
 import tech.taoq.system.mapper.DictTypeMapper;
 import tech.taoq.system.service.DictTypeService;
 
@@ -26,6 +27,10 @@ public class DictTypeServiceImpl implements DictTypeService {
     }
 
     public void deleteById(String id) {
+        DictTypeDO dictTypeDO = dictTypeMapper.selectById(id);
+        if (dictTypeDO.getInternal()) {
+            throw new ClientErrorException("内置记录不允许被删除");
+        }
         dictTypeMapper.deleteById(id);
     }
